Bug 78530 - Undoing message switches view from conversation to message if reading pane set OFF
Now I'm not able to reproduce the issue described in the bug originally. Tried various combination of steps with various view modes selected on reload.

I do see the issue with the undeleted message/conversation not appearing in the conv list after "undo" if page is refreshed in conv mode.

When we recreate a message after the deletion if it's not in cache we need to pass a list so that the list gets notified when message is moved. Otherwise the ui is not refreshed and message doesn't appear to be put back in the folder.

http://bugzilla.zimbra.com/show_bug.cgi?id=78530

Copied from Perforce
 Change: 505996