Bug 76253
Forwarding a mail with attached another email will lose its attachments using default options in preferences


The problem was that "hasAttach" was not set on the msg, despite having the "attachment" array set.
I believe the real problem is that it was used at all, instead of "msg.attachments" to determine if there are attachments.
The reason is that "hasAttach" is set via the flags (flag "a"), but for an RFC822 attachment, we do not get the flags. (probably since flags don't make sense for the most part ... it makes no sense to have stuff like isRead, isDraft, Flagged, etc on an attachment).

At first I thought of setting hasAttach in ZmNewWindow if isRfc822 is true, which worked, but then I realized we should not use hasAttach for this - hasAttach is probably only to be used for showing the attachment icon in the list view where we don't have the full message with attachments set. This is why it works fine for viewing the message but gets lost when forwarding, since this mistake is only in ZmComposeView (and not say ZmMailMsgView).

http://bugzilla.zimbra.com/show_bug.cgi?id=76253

Copied from Perforce
 Change: 506238