Bug 88100
Contact disappearing while using a draft


Like I suspected, the de-duplication is on purpose, but mostly for reply/reply-all to not have duplicates.

This is extremely minor edge case and nothing bad really happens, but since I found the place and it's an easy fix, here it is.

http://bugzilla.zimbra.com/show_bug.cgi?id=88100

Copied from Perforce
 Change: 506567