bug 87525 - make the editor default to bordered tables
Create tables with a border of width one and no cell spacing. Although
ever so slightly less aesthetically pleasing, it's more intuitive for
tables to actually look like tables.

(Please note that removing these borders with the table properties
dialog and re-adding them doesn't work as expected; this is an issue
in TinyMCE itself, reproducible on the demo instances on tinymce.com.)

http://bugzilla.zimbra.com/show_bug.cgi?id=87525

Copied from Perforce
 Change: 506824