Bug 87597 - Sent message doesn't show sender/recipient header after sending email
When notification for conversation is sent, the server doesn't return "to" addresses. All addresses get filtered out in the for loop and origLen is 0. To work around, in case when there's no "to" addresses and we are in the outbound folder, I'm leaving all addresses in part1 array which will trick origLen > 0. Then part1 will get overwritten with "to" addresses from msg.getAddresses()

http://bugzilla.zimbra.com/show_bug.cgi?id=87597

Copied from Perforce
 Change: 507090