Bug 87404
Intermittent "No such message exists." exception after replying to message


This looks like a duplicate of bug 74148, and bug 74148 was resolved, by exiting early from _sendMsg if it's autosave but the _autoSaveTimer (an idle timer) was killed (this was due to fact that clearing the interval does not necessarily do it right away, depending on the VM, so bug 74148 added this safety net of checking what we know for sure that would happen in case of attempting to kill the timer, which is null out that _timer member).

However, I this is just a guess as to what might still be wrong. Bug 66997 might be the cause. It added a call to _initAutoSave in ZmComposeController.prototype._handleResponseSendMsg. This means that if the user clicked "send" and the message was successfully sent, the _autoSaveTimer would start again!  Of course it does not sound likely that something wrong would happen since right away we close the view (thus killing it again, I believe). But who knows, maybe if something is slow (client or server I don't know) it still does actually get fired (somehow 3 seconds of idle time passed before the timer is killed again) in which case an attempt to save the draft would happen and get the exception.

http://bugzilla.zimbra.com/show_bug.cgi?id=87404

Copied from Perforce
 Change: 507642