Bug 88160
"No such message exists." exception after removing attachment from the message and trying to delete it


The single msg controller (ZmMsgController) was not updated to have the new msg (since deleting an attachment creates a whole new message without the attachment). This solution is to see if the view _controller has a setMsg method (such as ZmMsgController has) and if so call it with the new message.

http://bugzilla.zimbra.com/show_bug.cgi?id=88160

Copied from Perforce
 Change: 507991