Bug 82064
Need more descriptive error when moving a folder with name clash


resurrect my check-in above (or rather the fix to it in bug 84959 that fixed a regression from it).

this special logic for the already contains folder with same name case, was removed for some reason in Bug 82545. Other bugs touched around it, including bug bug 53906 that returned some of the logic that was removed by bug 82545 but not the code I added for this bug. Some of those bugs are related to touch client it seems (probably to ZWC and touch client at the same time). So need to make sure this fix does not cause a regression in any of those bugs.

http://bugzilla.zimbra.com/show_bug.cgi?id=82064

Copied from Perforce
 Change: 508027