Bug 88248 - Wrong search result returned for appointment in Calendar app
Removed conditional logic for ZmId.VIEW_CAL_LIST. This was causing wrong results on initial search. This wasn't apparent in 8.0 because we used to preserve the cal view type, but in 8.5 always switch to list.
Removed ZmCalListView.prototype.searchRefresh since it's not used anywhere else.

Discovered another issue: in the search results if you change the date range, it doesn't pass the userQuery and returns all appointments. Now passing the query in ZmCalListView.prototype._search to apptCache.getApptSummaries(params).

http://bugzilla.zimbra.com/show_bug.cgi?id=88248

Copied from Perforce
 Change: 508064