Bug 80208
While composing email and selecting vCard signature then vCard doesn't show as an attachment.


I believe it's a regression from bug 77707, that broke something bug 52128 did. The problem was changing AjxTimedAction.scheduleAction to setTimeout lost the "this" context. (Bug 52128 made the automatic draft saved when adding a vcard attachment - which happens right away - be delayed in 500 miliseconds, to not conflict with attach contacts zimlet and briefcase attachments).

http://bugzilla.zimbra.com/show_bug.cgi?id=80208

Copied from Perforce
 Change: 508253