Bug 87790
Message delete from conversation does not update the message count of the conversation


The code was updating the count (piggy back on the size) column correctly for reading pane on the bottom (a.k.a. multi-column) but not for reading pane on the right. Fix is simple - update the From field for reading pane on the right, as it includes that count. (see ZmConvListView.prototype._getCellContents)

http://bugzilla.zimbra.com/show_bug.cgi?id=87790

Copied from Perforce
 Change: 508298