Bug 80792 - IE: "Invalid argument" after double clicking to message from saved search result
Changeset:
1. ZmObjectManager.js

A url in the given MIME is of the following format: 
"http://0:0:0:0:0:0:0:1%0:7070/service/soap/ContactActionRequest:1331608015326:9c4f5868c5b0b4f2"

IE is unable to parse this url correctly and the node.href lookup throws an "Invalid Argument" error.

The parsing code in now enclosed within a try block to catch and supress this error.

http://bugzilla.zimbra.com/show_bug.cgi?id=80792

Copied from Perforce
 Change: 508321