Bug 87714
Bottom of msg cut off


There was a discrepancy between the calculated height of the test div that we move everything to to calculate the height, and the height of the same elements when they are in the body of the iframe.

The reason was the div was created in the shell (the regular document) and not in the iframe, that shell had some css definitions that affected the elements, and those definitions were not in the iframe.

The very simple solution: The only reason we moved the stuff to a div is to be able to get the offsetHeight (though we use a slightly different way to get the same result, _getHeightFromComputedStyle) of the div. So my solution is to move the stuff to a div in the same iframe body. That way it gets the exact same style. And in fact we don't need to move it back to the body, just leave it in the div, since each element moved to the div is gone from directly under the body. It looks the same and everything is fine.

http://bugzilla.zimbra.com/show_bug.cgi?id=87714

Copied from Perforce
 Change: 508536