Bug: 76040
Calendar refresh is done based on searchInProgress flag. If a search request is already in progress, another search request will not be fired.

This issue is not reproducible in local machine. But it is easily reproducible in our dogfood server. Just click refresh button 4-5 times duplicate appointments will be displayed. 

http://bugzilla.zimbra.com/show_bug.cgi?id=76040

Copied from Perforce
 Change: 508572