Bug 88837
Second msg into conv does not update conv list


This only happens when the first message was already in the current list view. (not say if the first one was sent - which was the case of bug 87861).

Another very confusing bug. At first I thought I did something wrong in bug 87861, which I thought is somewhat similar, but now I think it's not related. However, while investigating this, I realized I did do something wrong in bug 87861. While we needed to check in the appCtxt cache for the conv instead of creating it - the addMsg is already done later in sortAndNotify (and in the correct order). The only thing was to make sure not to re-create the conv with only the new message.

Anyway, so back to this bug - here the issue is that in the promotion of single message virtual conv to a real conv with 2 messages - we did not update the size (count in case of multiple messages) field. This was not an issue in case of bug 87861, since there there was no conv line already (since it was only in sent) so it was created, and correctly. (note that setting the F_SIZE field to update, also causes in the case of 1 or 2 message, the F_EXPAND field to update, which is what we want).

And finally there was also an issue that once the size was set to the single message conv line (from the message - bug 87416 made that change) - it was using it in ZmConvListView.prototype._getCellContents instead of the count. So I reset it to null if numMsgs is not 1.

Hopefully this covers the investigation and solution. :) It seems to work now.

http://bugzilla.zimbra.com/show_bug.cgi?id=88837

Copied from Perforce
 Change: 509019