Bug 88916
Bug 88916

Script error (TypeError: this._actionEv is null) when click to "Actions -> New Filter" from new window


This is not directly related to new window, and happens also in main window. But it's always happening in new window. The reason is that this._actionEv is only set when right clicking on a list item. (in _listActionListener). There's no list in new window, so it's never set. But it's not set in main window too, until the user right clicks on a list item.

So basically, this is a regression from bug 60017 - this._actionEv is not always set. It's only set on right click on a bubble.

My fix is simply a null test for this._actionEv.

There's another bug closely related that I will open separately.

http://bugzilla.zimbra.com/show_bug.cgi?id=88916

Copied from Perforce
 Change: 509156