Bug 89168
"create filter" context can be wrong


basically a correction/continuation of the fix to bug 88916.

The only solution I could find is to make a separate operation for the bubble "create filter" operation. _actionEv is a state variable that stays set after the right click actions menu is open and closed, and it's hard to get rid of it, so we need a different way to differentiate the two cases. This solution is not too complicated and is pretty straight forward so I think it's OK.

http://bugzilla.zimbra.com/show_bug.cgi?id=89168

Copied from Perforce
 Change: 509220