Bug 86202 - JS error accepting briefcase share
The problem that's causing the error is that setStatusMsg was being called on the wrong context.

ZmBaseController.showSummary works with either parentAppCtxt or appCtxt based on showToastOnParentWindow argument which wasn't passed as true.

Traced back to ZmMailListController.prototype._removeActionMsg where I added closeChildWin: appCtxt.isChildWindow to the parameters passed to list.moveItems that later ends up being passed to ZmBaseController.showSummary as showToastOnParentWindow.

http://bugzilla.zimbra.com/show_bug.cgi?id=86202

Copied from Perforce
 Change: 509247