Bug 88323 - Selected task not highlighted in list view on modifying status or mark as completed
After discussion with Eran removed multiple re-selection of tasks.

------
Pass specific item to _setNextSelection so that correct task is selected.

In cases when you edit a task and then save it, the task row looses class "Row-selected" until you hit "tab". I looked into Dwt.delClass which should be responsible for it, but it's not called. I think "forceSelection" is a little hacky, but does the job.

http://bugzilla.zimbra.com/show_bug.cgi?id=88323

Copied from Perforce
 Change: 509258