Bug 89118
Bottom of msg cut off


Even more concise solution to this issue.

Always wrap the content (when using Iframe to display a message) in a div within the body. Then there's no need to move anything around to calculate height. Another advantage is that we can calculate the height always (that can be done even without this fix), i.e. also when the height is larger than 150. So dragging the sash adjusts the height smoothly, and consistently.

http://bugzilla.zimbra.com/show_bug.cgi?id=89118

Copied from Perforce
 Change: 509449