Bug 68490
accepted appointments within shared mailbox are added to random calendars


For mountpoint (remote) calendars (and also for msg folders that are mountpoints), we only have account IDs available, we don't have an account object in our accountList for those.

The code was wrongly getting the local account as default in case the account was not set, thus the code was confused to think the default calendar of the remote account was the default of THIS account. And vice versa (due to the message account being wrong). So I make sure for mountpoints, to get the account ID correctly. Then when we compare stuff, use just the account IDs (compare the one from the msg with the one from each calendar to determine which to show by default).

The problem was with both calendars (which are organizers) and the message folders (which are also organizers).

Anyway as for behavior after the fix - the default selected calendar would be the default calendar of the account the message belongs to.

http://bugzilla.zimbra.com/show_bug.cgi?id=68490

Copied from Perforce
 Change: 509534