Bug 88571
Folder pane in ZWC does not expand large enough to view all folders / subfolders


Talked with Josh about it - no real UX reason to limit the max width of the tree.

Since we expose it though in skin.properties (can be set by setting TreeMaxWidth in there, which is translated to become window.skin.hints.tree.maxWidth) - decided to keep it there in case some customers prefer to limit it, so I set the default limit in the code to 1000 and in the skin.properties to 84rem which is also about 1000. This is big enough to cover all reasonable end user cases.

http://bugzilla.zimbra.com/show_bug.cgi?id=88571

Copied from Perforce
 Change: 509580