Bug: 89596 - Wrong toast message "1 Item Copied" when try to delete read-only draft message
See  Bug 26103 "when deleting an item in a folder shared to us, save a copy in our own trash"
Fix: Only do the copy if the shared folder is not read only.  There will still be a Copied message if the folder  is shared with Manager or Admin rights.

http://bugzilla.zimbra.com/show_bug.cgi?id=89596

Copied from Perforce
 Change: 510172