bug 87782 - fix editor resizing (again)
The editor would repeatedly show and hide itself during initial
setup. It's much better to keep the editor visible during
rendering, as this allows it to load immediately, hopefully
getting the right size once and for all.

Comments suggest that the primary motivation for hiding the editor
during load was to fall back on text mode on failure. As TinyMCE
has wider browser support than Zimbra -- as an example, it
supports IE7 -- this shouldn't occur in practice. Furthermore, the
scenario is unlikely to actually work unless we actually manage to
reproduce and test it.

http://bugzilla.zimbra.com/show_bug.cgi?id=87782

Copied from Perforce
 Change: 510288