Bug: 87993 - Problematic event locks up ajax client in ZWC when the Calendar app is selected
The problem was that the user created a daily appt with a duration of 6 months, mistaking appt start and end to be the start and end of its recurrence duration over the year, not its duration during the day.  The UI ends up creating thousands of entries, which cause it to slow to a halt.

This fix limits the extent of the duration based on the recurrence chosen, keeping it simple (not trying to replicate the server recurrence logic in the UI).  
The limits applied are:
Recurrence	Max Duration
Daily:     	Up to 1 day.
Weekly:   	Up to 1 week.
Monthly: 	Up to 5 weeks
Yearly:    	Up to 366 days.

http://bugzilla.zimbra.com/show_bug.cgi?id=87993

Copied from Perforce
 Change: 510327