Bug 87613 - Drag and drop of recurring all day event changes from 1 day to 2 days
Changeset:
1. ZmCalViewController.js

Description:
Added a check to set the end date equal to start date while modifying an All day event using DND.

For modifying an all day event the server expects the start and end date to be the same. 
In DND scenario, for an All Day Event, due to the offset(diff between old slot and slot the appt is dragged to) the start and end dates were getting set for 2 different days. 
So modifying an All day event sent the start and end dates as new start date and the consecutive date, thus converting it into a 2 day event.

http://bugzilla.zimbra.com/show_bug.cgi?id=87613

Copied from Perforce
 Change: 510723