Bug: 89832 - Location choose option not shown in suggested time pane
This happens prior to checking location suggestions.  The location resource loading was disabled for time suggestions in CL 497792, to prevent an error when !GAL_ENABLED - so no location data was available for the time suggestion.  I've changed the load to read location resources when GAL_ENABLED, whether its a location or time suggestion.

http://bugzilla.zimbra.com/show_bug.cgi?id=89832

Copied from Perforce
 Change: 510869