Bug: 86055 - Can't select Location: while creating folder
This is in the class of bugs that only happens in production,
due to a bug in Sencha Touch when adding data while
instantiating an object. The root of the problem is that the
organizer list did not know it was in edit mode. Explicitly
set that on the list when it's created, instead of passing it
as additional config.

http://bugzilla.zimbra.com/show_bug.cgi?id=86055

Copied from Perforce
 Change: 510915