Bug 88694 - Run Saved Search dialog doesn't show saved searches
this._overview.appName is "ZmZimbraMail and ZmApp.SEARCH_TYPES["ZmZimbraMail"] is undefined. Therefore when you pass each search to _isAllowed it's returning false and causing an empty list for searches.

To distinguish between the apps when calling the move dialog I'm creating a hash that will store a dialog for each app.

In ZmSearchTreeController.prototype._getMoveParams we shouldn't override params.overviewId because the dialogs are now going to be unique per app. We also don't need to override params.treeIds because we don't want to show folders in the list. Then the function becomes unnecessary.

Found one other issue that I'll also address in this fix — when saving a search, the dialog lists all the searches, not just the ones from the same app. To fix this I need to get the app that called the search. So at the time of showing the search results I'm setting appCtxt.searchAppName = app.getName() which is "Mail", "Tasks", etc (it's different from appCtxt.getCurrentAppName, which is "Search" in that case) and then I'm using appCtxt.searchAppName to store dialogs in the hash _newSearchDialogs[this.searchAppName].

http://bugzilla.zimbra.com/show_bug.cgi?id=88694

Copied from Perforce
 Change: 511144