Bug: 89756 - Network service Error on accepting appointment.
Changeset:
1.) ZtMsgWriter.js

In case of user taking action on invite from Mail app, isSeries & ridZ are unavailable. This is because action taken on invite from Mail app is on entire series. Whereas if user takes action on invite from Calendar Invite view, the action can be taken on series, instance or exception.

2.) ZtCalendarController.js
Passing isCalApp to help know the writer from which app the request was made.

http://bugzilla.zimbra.com/show_bug.cgi?id=89756

Copied from Perforce
 Change: 511377