Bug 89560
Notification handling: no 'expand' icon for new conversation after clicking "get mail"


First of all, this happens not only when clicking "get mail", but by any update (for example for my case, when clicking on a different item, it's enough that the update gets there from the server, in any way).

In bug 87790 I tried to make a simpler call, to just do updateField (to the ZmItem.F_EXPAND field) instead of redrawing the line. But that field was not there to begin with, so update didn't update it (or create it). That change did work for that case of delete, but this here bug is very similar to bug 88837, and not sure how I let this problem pass there. I don't see how that expand icon appeared. So I guess it didn't (just the count was updated after that fix).

So now I'm going back to doing redraw, even for switching from > 1 msg conv to 1, since otherwise it was not so consistent and left a blank area in place of the expand arrow. Also another small thing I fixed is to use the number of displayed messages in ZmConvListView.prototype._getAbridgedContent too to determine if it's a "conv" ( > 1 messages).

http://bugzilla.zimbra.com/show_bug.cgi?id=89560

Copied from Perforce
 Change: 512014