Bug 88507
Checkbox for selected item is automatically checked


Finally tracked down what was really going on.

First, note that bug 87418 is the same as this one, what Vince fix did was just add a feedback to the user that the automatically selected item (for view) is also checked (i.e. would be included in actions).

However, that did not really solve the issue of people not expecting that item to be included when checking others.

I checked on an 8.0 build and found that it indeed worked there (as far as not including the viewed item when checking others) - what it did was unselect (remove from selected items) the current viewed item, if it's not CHECKED.

The reason it was broken was that the DOM changed at some point, and somehow there are now 2 divs, the image one is inside the one we checked for the class, so it didn't work. (Maybe bug 84685 and bug 84554 is related to this, but doesn't matter).

My fix is actually to check whether the item is checked by using the item data instead of by comparing classes - safer and more direct and elegant anyway.

So for this fix I did that, and also undid the part of bug 97418 that checked the selected item. So now we are all back to the old way that's familiar to users.

http://bugzilla.zimbra.com/show_bug.cgi?id=88507

Copied from Perforce
 Change: 512027