Bug 79263
"Add To Contacts" option on RSS folder displayed "Loading..."


New related use-case added to this bug - right clicking on the "from" bubble in an RSS feed message (the one previously described was right clicking on the "from" in the list pane).

So fixed this similarly, simply return and not show the participant menu if there's no participant (no email address).

http://bugzilla.zimbra.com/show_bug.cgi?id=79263

Copied from Perforce
 Change: 512098