Bug 89968
Msg added to conv not sized correctly


The problem was _usingIframe was not set when _resize was called, it was a timing issue. It was not visible on localhost since it's faster. However I found a solution that relates to the idea we need to do the same thing for Conv view and Msg view. This is very partial - the MSG view already called a resize (_resetIframeHeight), but that was only working for taller than 150px. So what I do is use that call to call the capsule view resize instead at that same point (which is at the end of ZmMailMsgView.prototype._displayContent, in which case we already have _usingIframe set correctly), and also at other points... ZmMailMsgView._resetIframeHeight is no longer called in those cases, but it didn't make sense anyway since ZmMailMsgCapsuleView.prototype._resize would anyway override what _resetIframeHeight did or worse - there would be a race condition. And I think it has lots of code that's no longer relevant anyway, definitely not for conv view case.
The actual code fix is to rename ZmMailMsgCapsuleView.prototype._scheduleResize to ZmMailMsgCapsuleView.prototype._resetIframeHeightOnTimer, so it overrides the one from ZmMailMsgView which is called correctly.

http://bugzilla.zimbra.com/show_bug.cgi?id=89968

Copied from Perforce
 Change: 512115