Bug 90674
Clicking to search folder switches to different view rather than saving previous state


Bug 89392 caused a regression of bug 85137. Since both were verified already, I create this clone to fix the issue.

Fix is to have an explicit param forceTypes if we want to not override the type (msg/conv) in case of mail to the one the user pref is currently set for (implicit pref of msg vs. conv). This is set in ZmProgressController.prototype._nextChunk to true (along with the type set to ZmItem.MSG). Any other cases like that could use the param as well.

http://bugzilla.zimbra.com/show_bug.cgi?id=90674

Copied from Perforce
 Change: 512174