Bug 90566
Selecting a conversation triggers mailbox server DosFilter


Bug 73711 added that call, and according to the comment it's for the case of delayed "mark read". So I added this condition to only call it in the case of delayed mark read.

ZmConvListController.prototype._displayItem is also preventing single MsgActionRequests from firing by setting waitOnMarkRead to true. But even after it's returning from the call to this._doublePaneView.setItem(item); and setting waitOnMarkRead - the messages in the conv still have isUnread true. And that's why (without my fix to _handleResponseLoadMessage1), ZmMailListController.prototype._handleMarkRead is executing the block inside the if (item && item.isUnread && !waitOnMarkRead) and thus causing a MsgActionRequest to mark the message as "read", for each message.

This might be due to timing of the notifications that are eventually causing the messages to have isUnread false.

http://bugzilla.zimbra.com/show_bug.cgi?id=90566

Copied from Perforce
 Change: 512434