bug 89920, 90004 - editor: allow accessing browser native context menu
This change allows accessing the builtin context menu on
Ctrl+RightClick, and uses the document way of enabling the browser's
native spell checker.

There's been a bit of back and forth on these features:

The browser spell check was originally enabled in bug 46098 and 44103,
primarily for Zimbra Desktop. AFAICT Zimbra rolled its own editor at
the time, and I assume it used the usual Zimbra context menu. In order
to work around this, a control-right click was made to trigger the
native browser context menu.

By comparison, although Zimbra originally used the TinyMCE context
menu, it was disabled as part of bug 51671. I re-enabled it as part of
the upgrade to TinyMCE 4, but disabled it in bug 85153 due to issues
with Zimbra's spell checker. I then later resolved these issues, and
re-enabled them in bug 81955. But the Ctrl-RightClick trick never
worked with TinyMCE 4. This change addresses that.

http://bugzilla.zimbra.com/show_bug.cgi?id=89920
http://bugzilla.zimbra.com/show_bug.cgi?id=90004

Copied from Perforce
 Change: 512725