Bug 55760
Status in mobile device tab in user preference does not updates


Can't reproduce the actual bug (explanation of what they might see is in the bug).

Fix to minor issue found when researching this bug:
Reset the operations when loading the devices, and also disable the "wipe" button if wipe is completed.

http://bugzilla.zimbra.com/show_bug.cgi?id=55760

Copied from Perforce
 Change: 512917