Bug 91083
no 'expand' icon after replying to a single message conversation


The continuous saga of this very tricky code. Once again, post the change of filtering out trashed messages from conversation view, things are way more complicated. Originally to know if an "expand" icon is needed, we relied on numMsgs (the total number of messages in the conv), but now we rely on knowing the specific folders of each message in the conv (which is in the msgFolder associative array of ZmConv). The problem here was the timing - msgFolder is still not updated when ZmConvListView.prototype._changeListener calls _getDisplayedMsgCount  (numMsgs is already 2, but msgFolder is still just for one message).  msgFolder is updated in addMsg but it is called after this particular call to ZmConvListView.prototype._changeListener  is called (with ZmEvent.E_MODIFY event and fields[ZmItem.F_SIZE] set to true.
I tried to fix it by setting msgFolder in places like when setting fields[ZmItem.F_SIZE] = true in ZmMailList.prototype.notifyCreate, but things didn't get called when I thought they would etc.
So at the end, I added another slightly hacky thing another case where we force the update of the size - this time whenever a message is created and added to a conv. The same as I already did when moving a message to or from the conversation.

http://bugzilla.zimbra.com/show_bug.cgi?id=91083

Copied from Perforce
 Change: 513025