Bug: 90854 - Quick reply quotes first message in conv, not latest
Change how we find the ‘original msg’ in a conv to use for
replies and forwards: the most recent msg that’s not in
Trash/Junk/Sent (unless the user is in one of those). Looking
for the most recent msg doesn’t depend on conv ordering (old
to new or new to old). Look in the msg store rather than the
conv’s separate list of messages. Stop maintaining that parallel
list of messages.

Each conv returned by the server within search results has
a list of msgs with minimal info (ID and folder). Instead of
instantiating each one of those as a ZtMailMsg, keep it as
JSON since all we really need is the folder ID. Maybe we’ll
even get a performance gain as it should reduce our memory use.

http://bugzilla.zimbra.com/show_bug.cgi?id=90854

Copied from Perforce
 Change: 513052