bug 81955, 89920, 90004 - ruthlessly kill the TinyMCE context menu
The TinyMCE context adds very little extra functionality to our
editor, yet it actively confuses users. When they right click on an
edit field, they expect the usual menu with copy/paste/whathaveyou. In
addition, the built-in spell checker relies on it, and it works much
better than Zimbra's spell checker.

The only reason I reintroduced it was as possible fix for bug 81955,
but that's wrong: the contextmenu itself is much more useful than the
(new) ability to attach images by URL. In addition, we shouldn't ship
8.5 with this menu enabled only to disable it later.

The final nail in the coffin is that Ctrl+RightClick for triggering
the native context menu is awfully unintuitive.

(The event-location-swizzling in _handleEditorEvent() was added in
changelist 221391 as part of spell checking support. It appears
unnecessary for its original purpose, yet now breaks the native
context menu.)

http://bugzilla.zimbra.com/show_bug.cgi?id=81955
http://bugzilla.zimbra.com/show_bug.cgi?id=89920
http://bugzilla.zimbra.com/show_bug.cgi?id=90004

Copied from Perforce
 Change: 513141