bug 89822 - adapt appt. editor to changes in attendee height
Adding and removing attendees can change the height and width of the
field. That's subtly wrong, but hard to address due to intricacies of
how the field and its bubbles are laid out. To work around this, we
ensure that the appointment editor does a full relayout of its
contents by calling resize().

This also fixes the logic for showing and hiding the scheduler: Even
though the comment suggests that the code merely "updates the schedule
view even if it won't be visible" it actually does show it, but in a
buggy fashion that fails to toggle the show/hide button. Instead, we
just force it open using _toggleInlineScheduler().

http://bugzilla.zimbra.com/show_bug.cgi?id=89822

Copied from Perforce
 Change: 513551