Bug 88986 - DL doesn't expand from opened appointment UI
Passing canExpand to new AjxEmailAddress where possible instead of assigning after an object is created.
I did this in the places where it's absolutely straightforward. There's an occurence in ZmContactsHelper that I left alone. There's conditional logic after AjxEmailAddress object is created before isGroup and canExpand are assigned -- I don't think it's worth the risk to mess with it at this point.

====

When the AjxEmailAddress is created in _getAttendeesByRoleCollapsed pass isGroup() and canExpand. The lack of them causes autocomplete list view not to be initialized in ZmMailMsgView.prototype._getBubbleHtml

http://bugzilla.zimbra.com/show_bug.cgi?id=88986

Copied from Perforce
 Change: 513586