Bug 89554
List selection can be confusing, with or without checkboxes


Implemented another part of this - in bug 91013 Josh added a style for "alt selected" - I use it here for marking currently viewed item in the list, if the item is viewed but is not selected for action. Also implemented some extra behavior around this.
Two main cases:
(1)
a. User selects an item for view.
b. User then right clicks on a different item - as the actions menu pops up, the "viewed but unselected" style (lighter blue) is applied for the viewed item in the list.
c. When user is done with the menu (dismissing it explicitly or by performing an action) - new functionality is to re-select the viewed item. (and reseting the toolbar actions accordingly)

(2)
a. User selects an item for view
b. User then clicks on a checkbox (only if checkboxes are on of course) of a different item.
c. As a result, the item from (a) is marked with the "viewed but unselected" style. (in addition to the "number of convs selected" being updated to 1, per Conrad's previous fix
d. If the user then unchecks the checkbox from (b) (or multiple checkboxes similarly checked after step b) - the viewed item remains in that "viewed and unselected" state, and the toolbar buttons are all disabled. This is fine, I confirmed with Josh too. The user can click on that item again to select it for action.

There's also a 3rd case actually:
(3)
a. User selects multiple items (using checkboxes, or shift click), including an item for view
b. User right clicks on an item outside of selection.
c. The viewed item is marked as above.
d. When the popup is popped down, only the viewed item is selected, not the multiple ones from step a. I believe this is good enough for this edge of an edge case. Josh agreed too.

I believe this also solves the issue in Bug 89834, that Zhenia worked on.

A couple changes I did in the code that might need explanation:
(A) in a few places I simplified a removal of style to remove the style itself, instead of all the "select" styles - it's hard to explain but I think it makes more sense and is simpler. I don't think there was a good reason for the previous code. Good example is for the cases of the keyboard anchor highlighting. Some code was checking also if something is selected by doing string indexOf for the classes, not the best way, and I either replaced some of those with more direct ways to know or with simplified code. I did keep some of those, which brings me to -
(B) I was able to get it all to work (except for one edge case, doesn't matter what exactly) despite my concern that we rely on regex and string comparisons for stuff like "Row-selected", and the new class is "Row-selected-alt" - which could and does confuse the code in some cases (now more rare due to my changes in (A)). So I decided to change the class name (from bug 91013) from Row-selected-alt to Row-altSelected, and that fixed even the edge case.

http://bugzilla.zimbra.com/show_bug.cgi?id=89554

Copied from Perforce
 Change: 513591