Bug: 85559 - IE: Multiple blank lines shown up for HTML mail
The messages contain external images, with height and width specified in the html; when they are replaced with the 1x1-trans.png image upon initial render, the browser either:
- honors the original height specification (this is happending with sample mail 2) or
- Sets the image height to equal the width.

I'm not sure why, and could spend a day tracking this down.  So a simple fix (which I would maybe have to apply anyway) is to set the transparent 1x1 display to none, and then reset it to block when the user clicks display images.  That provides the proper hiding/display of the images.

- The display looks fine in Chrome (it was broken there too before the fix); 
- In IE most of the spacing is gone, but the message itself has html that is setting the height for various table cells, which apparently IE honors even though the content of the cell is an undisplayed image (display==none).  For that the only fix is to change the email html, or use another browser.
- In Firefox, there are some spacing issues too - these seem to stem from the user email using a table with height=1005 !important - it sizes properly when that is disabled, or if the table cell uses vertical-align top.

http://bugzilla.zimbra.com/show_bug.cgi?id=85559

Copied from Perforce
 Change: 513742