Bug: 90692
Expires directive is not set for offline cookie which made it as a session cookie. Once the user closes ZWC offline cookie is deleted. Now expires directive is set for offline cookie as 1 month, which makes the subsequent login successful. 

Separator is replaced from “,” to “_” to avoid the decoding in the server side.

http://bugzilla.zimbra.com/show_bug.cgi?id=90692

Copied from Perforce
 Change: 513925