bug 88829 - touch: ensure logins persist across sessions
It appears 'loginTouch.jsp' doesn't handle the actual login, but has
its form submit to the regular 'login.jsp'. Whereas the former
manually forced the 'rememberme' argument to zm:login to true, the
latter sets it based on the zrememberme parameter -- which the former
neglects to set. The simplest solution to this is adding 'zrememberme'
as a hidden parameter to the touch login form.

http://bugzilla.zimbra.com/show_bug.cgi?id=88829

Copied from Perforce
 Change: 513986