bug 87782 - fix editor resizing during load on Firefox
For some reason or other, the TinyMCE toolbar is detected as unreasonably
large during editor load. My best guess is that this is race condition
between JavaScript and CSS; the toolbars have been created, but the
CSS for them isn't in yet.

(We really should migrate all this resize logic into TinyMCE; we can't
possibly be the only ones wanting the resize it dynamically.)

http://bugzilla.zimbra.com/show_bug.cgi?id=87782

Copied from Perforce
 Change: 514029